Fixed where the ball was spawning by changing its radius size from 7, to 0.25L as stated within the specification.

+++ src/model/Board.java
@@ -45,7 +45,7 @@ public interface Board {
 
     public void setBallSpeed(String ballId, int velocityX, int velocityY);
 
-	public String addBall(String string, int x, int y);
+	public String addBall(String string, double x, double y);
 
 	public double getGravity();
